Move the real pebble tests to a separate module.
